refactor as module
